Extract tty.h and tty.c
