+++ src/Makefile.am
@@ -2,5 +2,5 @@ AUTOMAKE_OPTIONS=subdir-objects
 AM_CFLAGS=-Wall -Wextra -pedantic-errors -Wno-unused-parameter -Werror
 bin_PROGRAMS=pick
 dist_pick_SOURCES=choice.c choice.h choices.c choices.h io.c io.h main.c \
-				  ui.c ui.h compat/queue.h compat/strlcpy.c compat/strlcpy.h \
-				  compat/strlcat.c compat/strlcat.h
+				  tty.c tty.h ui.c ui.h compat/queue.h compat/strlcpy.c \
+				  compat/strlcpy.h compat/strlcat.c compat/strlcat.h
